+++ src/Planners/src/UnicycleTrajectoryPlanner.cpp
@@ -16,6 +16,7 @@
 #include <BipedalLocomotion/Planners/UnicycleUtilities.h>
 #include <BipedalLocomotion/TextLogging/Logger.h>
 
+#include <FeetCubicSplineGenerator.h>
 #include <iDynTree/KinDynComputations.h>
 #include <iDynTree/Model.h>
 
@@ -85,6 +86,19 @@ public:
     };
 
     COMHeightTrajectory comHeightTrajectory;
+
+    struct FootTrajectory
+    {
+
+        std::vector<iDynTree::Transform> transform; /**< Foot transform */
+        std::vector<iDynTree::Twist> mixedVelocity; /**< Spatial velocity in mixed
+                                                           representation */
+        std::vector<iDynTree::SpatialAcc> mixedAcceleration; /**< Spatial acceleration in mixed
+                                                       representation */
+    };
+
+    FootTrajectory rightFootTrajectory;
+    FootTrajectory leftFootTrajectory;
 };
 
 BipedalLocomotion::Planners::UnicycleTrajectoryPlannerInput BipedalLocomotion::Planners::
@@ -180,6 +194,16 @@ bool Planners::UnicycleTrajectoryPlanner::setRobotContactFrames(const iDynTree::
     }
 
     return true;
+}
+
+int Planners::UnicycleTrajectoryPlanner::getLeftContactFrameIndex() const
+{
+    return m_pImpl->parameters.leftContactFrameIndex;
+}
+
+int Planners::UnicycleTrajectoryPlanner::getRightContactFrameIndex() const
+{
+    return m_pImpl->parameters.rightContactFrameIndex;
 }
 
 bool Planners::UnicycleTrajectoryPlanner::initialize(
@@ -284,6 +308,8 @@ bool Planners::UnicycleTrajectoryPlanner::initialize(
     ok = ok && loadParamWithFallback("nominalWidth", m_pImpl->parameters.nominalWidth, 0.20);
     ok = ok && loadParamWithFallback("minWidth", minWidth, 0.14);
     ok = ok && loadParamWithFallback("minStepDuration", minStepDuration, 0.65);
+    m_pImpl->parameters.minStepDuration
+        = std::chrono::nanoseconds(static_cast<int64_t>(minStepDuration * 1e9));
     ok = ok && loadParamWithFallback("maxStepDuration", maxStepDuration, 1.5);
     ok = ok && loadParamWithFallback("nominalDuration", nominalDuration, 0.8);
     ok = ok && loadParamWithFallback("maxAngleVariation", maxAngleVariation, 18.0);
@@ -381,6 +407,23 @@ bool Planners::UnicycleTrajectoryPlanner::initialize(
     // Set the dynamical system to the integrator
     ok = ok && m_pImpl->comSystem.integrator->setDynamicalSystem(m_pImpl->comSystem.dynamics);
     ok = ok && m_pImpl->comSystem.integrator->setIntegrationStep(m_pImpl->parameters.dt);
+
+    // initialize the feet cubic spline generator
+    auto feetCubicSplineGenerator = m_pImpl->generator.addFeetCubicSplineGenerator();
+
+    double stepHeight;
+    double stepLandingVelocity;
+    double footApexTime;
+    double pitchDelta;
+    ok = ok && loadParamWithFallback("stepHeight", stepHeight, 0.05);
+    ok = ok && loadParamWithFallback("stepLandingVelocity", stepLandingVelocity, -0.1);
+    ok = ok && loadParamWithFallback("footApexTime", footApexTime, 0.5);
+    ok = ok && loadParamWithFallback("pitchDelta", pitchDelta, 0.0);
+
+    feetCubicSplineGenerator->setFootApexTime(footApexTime);
+    feetCubicSplineGenerator->setFootLandingVelocity(stepLandingVelocity);
+    feetCubicSplineGenerator->setPitchDelta(pitchDelta);
+    feetCubicSplineGenerator->setStepHeight(stepHeight);
 
     // generateFirstTrajectory;
     ok = ok && generateFirstTrajectory();
@@ -616,8 +659,10 @@ bool Planners::UnicycleTrajectoryPlanner::advance()
         return outputVect;
     };
 
-    m_pImpl->output.dcmTrajectory.position = convertToEigen(dcmGenerator->getDCMPosition());
-    m_pImpl->output.dcmTrajectory.velocity = convertToEigen(dcmGenerator->getDCMVelocity());
+    Planners::UnicycleUtilities::Conversions::convertVector(dcmGenerator->getDCMPosition(),
+                                                            m_pImpl->output.dcmTrajectory.position);
+    Planners::UnicycleUtilities::Conversions::convertVector(dcmGenerator->getDCMVelocity(),
+                                                            m_pImpl->output.dcmTrajectory.velocity);
 
     // get the CoM planar trajectory
     std::chrono::nanoseconds time = m_pImpl->input.initTime;
@@ -677,6 +722,42 @@ bool Planners::UnicycleTrajectoryPlanner::advance()
         m_pImpl->output.comTrajectory.acceleration[i].z()
             = m_pImpl->comHeightTrajectory.acceleration[i];
     }
+
+    // get the feet trajectories
+    auto feetCubicSplineGenerator = m_pImpl->generator.addFeetCubicSplineGenerator();
+
+    feetCubicSplineGenerator->getFeetTrajectories(m_pImpl->leftFootTrajectory.transform,
+                                                  m_pImpl->rightFootTrajectory.transform);
+    feetCubicSplineGenerator
+        ->getFeetTwistsInMixedRepresentation(m_pImpl->leftFootTrajectory.mixedVelocity,
+                                             m_pImpl->rightFootTrajectory.mixedVelocity);
+    feetCubicSplineGenerator
+        ->getFeetAccelerationInMixedRepresentation(m_pImpl->leftFootTrajectory.mixedAcceleration,
+                                                   m_pImpl->rightFootTrajectory.mixedAcceleration);
+
+    Planners::UnicycleUtilities::Conversions::convertVector(m_pImpl->leftFootTrajectory.transform,
+                                                            m_pImpl->output.leftFootTrajectory
+                                                                .transform);
+    Planners::UnicycleUtilities::Conversions::convertVector(m_pImpl->rightFootTrajectory.transform,
+                                                            m_pImpl->output.rightFootTrajectory
+                                                                .transform);
+
+    Planners::UnicycleUtilities::Conversions::convertVector(m_pImpl->leftFootTrajectory
+                                                                .mixedVelocity,
+                                                            m_pImpl->output.leftFootTrajectory
+                                                                .mixedVelocity);
+    Planners::UnicycleUtilities::Conversions::convertVector(m_pImpl->rightFootTrajectory
+                                                                .mixedVelocity,
+                                                            m_pImpl->output.rightFootTrajectory
+                                                                .mixedVelocity);
+    Planners::UnicycleUtilities::Conversions::convertVector(m_pImpl->leftFootTrajectory
+                                                                .mixedAcceleration,
+                                                            m_pImpl->output.leftFootTrajectory
+                                                                .mixedAcceleration);
+    Planners::UnicycleUtilities::Conversions::convertVector(m_pImpl->rightFootTrajectory
+                                                                .mixedAcceleration,
+                                                            m_pImpl->output.rightFootTrajectory
+                                                                .mixedAcceleration);
 
     // get the merge points
     m_pImpl->generator.getMergePoints(m_pImpl->output.mergePoints);
@@ -798,3 +879,9 @@ BipedalLocomotion::Planners::UnicycleTrajectoryPlanner::getContactPhaseList()
 
     return contactPhaseList;
 };
+
+std::chrono::nanoseconds
+BipedalLocomotion::Planners::UnicycleTrajectoryPlanner::getMinStepDuration() const
+{
+    return m_pImpl->parameters.minStepDuration;
+}